requirements.txt
absl-py==1.4.0
astunparse==1.6.3
cachetools==5.2.1
charset-normalizer==3.0.1
# contourpy==1.0.7
cycler==0.11.0
fonttools==4.38.0
gast==0.3.3
google-auth==2.16.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.51.1
# h5py==2.10.0
idna==3.4
importlib-metadata==6.0.0
# Keras==2.3.0
# Keras-Applications==1.0.8
# Keras-Preprocessing==1.1.2
# keras-vggface==0.6
kiwisolver==1.4.4
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib
numpy
oauthlib==3.2.2
# opt-einsum==3.3.0
packaging==23.0
Pillow==9.4.0
protobuf==3.20.3
psutil==5.9.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==3.0.9
python-dateutil==2.8.2
PyYAML==6.0
requests==2.28.2
requests-oauthlib==1.3.1
rsa==4.9
scipy
six==1.16.0
tensorboard
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow-estimator==2.3.0
termcolor==2.2.0
urllib3==1.26.14
Werkzeug==2.2.2
wrapt==1.14.1
zipp==3.11.0
click==8.0.4
transformers==4.21.1
opencv-python==4.6.0.66
datasets==2.8.0
tensorflow
ultralytics==8.0.100
seaborn
soundfile
librosa
torchvision
sentencepiece
torchaudio
gurobipy
tabulate
kubernetes
yq
black
black[jupyter]
jiwer
statsmodels
soundfile
librosa
sklearn